Changed name to MessagePane, moved fetch to MessagePane
